Update authors.html
I changed the format of the pictures to just be a table, and the text on bottom and top are rows in the table.
I found a better contact form and used that code and copied the CSS at the top. Originally, I wanted the contact form to extend across the page, but I don't think it would look good. I think it should just be in the footer. We should copy this code for each of the other pages. What I still need to do:

Create a footer (add more to it with links to homepage and to social media)
figure out the twitter timeline shit
get resume from amina, upload mine