update authors.html
I changed the footer so it has a table in it. The contact form wouldn't html nicely into the table, so I went ahead and just made a table for the other two sections.

I cannot figure out that damned twitter timeline. Fuck that shit.

Fixed the bios and resumes, so that all should be correct.